modificacion en eliminar usuario y eliminar documento

+++ back/controllers/documentos.js
@@ -1,5 +1,6 @@
 const db = require('../config');
 const path = require('path');
+const fs = require('fs');
 
 const getDocumentos = (req, res) => {
     const query = 'SELECT * FROM documentos';
@@ -14,9 +15,7 @@ const getDocumentos = (req, res) => {
 };
 
 const uploadDocumento = (req, res) => {
-
     const documento = req.files.documento;
-
     const uploadPath = path.join(__dirname, '../uploads', documento.name);
 
     documento.mv(uploadPath, (err) => {
@@ -25,8 +24,8 @@ const uploadDocumento = (req, res) => {
             return res.status(500).send(err);
         }
 
-        const query = 'INSERT INTO documentos ( documento) VALUES (?)';
-        db.query(query, [ documento.name], (err, result) => {
+        const query = 'INSERT INTO documentos (documento) VALUES (?)';
+        db.query(query, [documento.name], (err, result) => {
             if (err) {
                 console.error('Error al guardar documento en la base de datos:', err);
                 return res.status(500).send(err);
@@ -38,13 +37,44 @@ const uploadDocumento = (req, res) => {
 
 const deleteDocumento = (req, res) => {
     const { id } = req.params;
-    const query = 'DELETE FROM documentos WHERE documentosid = ?';
-    db.query(query, [id], (err, result) => {
+
+    
+    const getQuery = 'SELECT documento FROM documentos WHERE documentosid = ?';
+    db.query(getQuery, [id], (err, results) => {
         if (err) {
-            console.error('Error al eliminar documento:', err);
+            console.error('Error al obtener documento:', err);
             return res.status(500).send(err);
         }
-        res.status(200).send('Documento eliminado correctamente');
+
+        if (results.length === 0) {
+            return res.status(404).send('Documento no encontrado');
+        }
+
+        const documentoNombre = results[0].documento;
+        const filePath = path.join(__dirname, '../uploads', documentoNombre);
+
+        
+
+
+        const deleteQuery = 'DELETE FROM documentos WHERE documentosid = ?';
+        db.query(deleteQuery, [id], (err, result) => {
+            if (err) {
+                console.error('Error al eliminar documento:', err);
+                return res.status(500).send(err);
+            }
+
+            
+
+            
+            fs.unlink(filePath, (err) => {
+                if (err) {
+                    console.error('Error al eliminar archivo:', err);
+                    return res.status(500).send(err);
+                }
+
+                res.status(200).send('Documento eliminado correctamente');
+            });
+        });
     });
 };
